+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - (Bugfix) Orphan PVC are not removed
 - (Bugfix) Remove LocalStorage Deadlock
 - (Bugfix) Skip arangosync members state inspection checks
+- (Feature) Add LocalStorage DaemonSet Priority support
 
 ## [1.2.10](https://github.com/arangodb/kube-arangodb/tree/1.2.10) (2022-04-27)
 - (Feature) Allow configuration for securityContext.runAsUser value